+++ test/dummy/app/helpers/application_helper.rb
@@ -1,9 +1,2 @@
 module ApplicationHelper
-  def dummy_page_nav(title:, back_url: nil, back_label: "Home")
-    recording_studio_page_nav(
-      title: title,
-      page_nav_back_url: back_url,
-      page_nav_back_label: back_label
-    )
-  end
 end