+++ test/dummy/config/initializers/recording_studio_publishable.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+RecordingStudioPublications::FamilyManagement.install!